Added stuff for model evaluation and metrics.
This included restructuring model-test.ipynb a bit to seperate out the
data preprocessing (window generation and such) and put it in
final_data_processing.ipynb.

Given a data file such as X_top300_201701_201912_30_m.npy, the way to
proceed is now to run final-data_processing.ipynb, then
model-test.ipynb (for whatever models we want to run), then
evaluate_models.ipynb.

Also added processed_data and stored_models folders to gitignore, since
the contents of these folders can get quite large. These materials can
be shared through other means, or generated on one's own machine as
needed.

Note that in evaluate_models.ipynb, a couple of the plots generated
might not quite be working the way they are supposed to. The plot
generation code is derived from the old evaluate_multiply.ipynb (for the
traffic data), and I wasn't entirely sure in some places how to properly adapt them to
the new data. That is still a TODO.